Make project compile on systems without cgo.
Otherwise fails with a "build constraints exclude all Go files in" error
from opengl.
